feat(institution): add delete and details modals
